ocr_defense/attacks/adv_docvqa/requirements.txt
sentencepiece
transformers
torch
torchvision
secml-torch
protobuf
pandas
editdistance
